Centralize the web-core
