+++ packages/web-core/package.json
@@ -46,7 +46,9 @@
     "@mui/material": "^5.10.17",
     "@mui/icons-material": "^5.0.0",
     "@jbrowse/product-core": "^2.5.0",
+    "@jbrowse/app-core": "^2.5.0",
     "copy-to-clipboard": "^3.3.1",
+    "clone": "^2.0.0",
     "react-error-boundary": "^4.0.3",
     "shortid": "^2.2.15"
   },
