+++ packages/web-core/src/BaseWebSession/AboutDialog.tsx
@@ -1,0 +1,1 @@
+export { AboutDialog as default } from '@jbrowse/product-core'
