+++ packages/web-core/src/index.ts
@@ -1,1 +1,2 @@
 export * from './SessionConnections'
+export * from './BaseWebSession'
