+++ products/jbrowse-desktop/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "../../packages/product-core/src/Session/Themes.ts"],
+  "include": ["src"],
   "compilerOptions": {
     "jsx": "react-jsx",
     "noEmit": true
