+++ products/jbrowse-react-app/package.json
@@ -110,7 +110,5 @@
   "peerDependencies": {
     "react": ">=16.8.0"
   },
-  "publishConfig": {
-    "access": "public"
-  }
+  "private": true
 }
