+++ products/jbrowse-react-app/src/index.ts
