+++ products/jbrowse-react-app/tsconfig.build.es5.json
@@ -12,8 +12,7 @@
   "exclude": [
     "src/**/*.test.ts*",
     "src/**/*.test.js*",
-    "src/makeWorkerInstance.ts",
-    "src/indexESM.ts"
+    "src/makeWorkerInstance.ts"
   ],
   "references": [
     { "path": "../../packages/core/tsconfig.build.json" },
@@ -25,6 +24,7 @@
     { "path": "../../plugins/linear-genome-view/tsconfig.build.es5.json" },
     { "path": "../../plugins/sequence/tsconfig.build.es5.json" },
     { "path": "../../plugins/svg/tsconfig.build.es5.json" },
+    { "path": "../../plugins/hic/tsconfig.build.es5.json" },
     { "path": "../../plugins/trix/tsconfig.build.es5.json" },
     { "path": "../../plugins/variants/tsconfig.build.es5.json" },
     { "path": "../../plugins/wiggle/tsconfig.build.es5.json" }
