+++ products/jbrowse-react-app/tsconfig.build.esm.json
@@ -8,7 +8,7 @@
     "composite": true
   },
   "include": ["./src/**/*.ts*", "./src/**/*.js*"],
-  "exclude": ["src/**/*.test.ts*", "src/**/*.test.js*", "src/indexES5.ts"],
+  "exclude": ["src/**/*.test.ts*", "src/**/*.test.js*"],
   "references": [
     { "path": "../../packages/core/tsconfig.build.json" },
     { "path": "../../plugins/alignments/tsconfig.build.esm.json" },
@@ -19,6 +19,7 @@
     { "path": "../../plugins/linear-genome-view/tsconfig.build.esm.json" },
     { "path": "../../plugins/sequence/tsconfig.build.esm.json" },
     { "path": "../../plugins/svg/tsconfig.build.esm.json" },
+    { "path": "../../plugins/hic/tsconfig.build.esm.json" },
     { "path": "../../plugins/trix/tsconfig.build.esm.json" },
     { "path": "../../plugins/variants/tsconfig.build.esm.json" },
     { "path": "../../plugins/wiggle/tsconfig.build.esm.json" }
